Initial split of compiler
