+++ BUILD.gn
@@ -117,6 +117,7 @@ ts_sources = [
 group("deno_deps") {
   deps = [
     ":msg_rs",
+    ":snapshot_compiler",
     ":snapshot_deno",
     "libdeno:libdeno_static_lib",
   ]
@@ -220,6 +221,30 @@ run_node("bundle") {
   ]
 }
 
+run_node("bundle_compiler") {
+  out_dir = "$target_gen_dir/bundle/"
+  outputs = [
+    out_dir + "main_compiler.js",
+    out_dir + "main_compiler.js.map",
+  ]
+  depfile = out_dir + "main_compiler.d"
+  deps = [
+    ":msg_ts",
+  ]
+  args = [
+    rebase_path("third_party/node_modules/rollup/bin/rollup", root_build_dir),
+    "-c",
+    rebase_path("rollup.config.js", root_build_dir),
+    "-i",
+    rebase_path("js/main_compiler.ts", root_build_dir),
+    "-o",
+    rebase_path(out_dir + "main_compiler.js", root_build_dir),
+    "--sourcemapFile",
+    rebase_path("."),
+    "--silent",
+  ]
+}
+
 ts_flatbuffer("msg_ts") {
   sources = [
     "src/msg.fbs",
@@ -237,5 +262,13 @@ snapshot("snapshot_deno") {
   source_root = "$target_gen_dir/bundle/main.js"
   deps = [
     ":bundle",
+  ]
+}
+
+# Generates $target_gen_dir/snapshot_compiler.bin
+snapshot("snapshot_compiler") {
+  source_root = "$target_gen_dir/bundle/main_compiler.js"
+  deps = [
+    ":bundle_compiler",
   ]
 }
